utils: Use Android version of DTC
Using Android version of DTC resulting more correct of dumped DTS. The
one that I notice is comma seperator between value. With Linux DTC, the
comma seperator is converted to '\0', while Android DTC convert these
normally.

I don't know the difference of those, but if there a normal version, why
not?

The source of Android DTC is taken from here:
https://android.googlesource.com/platform/external/dtc, branch master.